🔧 chore(config): update site title and URL for deployment
Changed the site title to 'Faizan Siddiqui' and updated the production
URL to 'https://faizansiddiqui.me' in the Docusaurus configuration.
Also modified the organization and project names to reflect the
correct GitHub repository details. This ensures that the site is
properly configured for deployment and branding.

Additionally, commented out the edit URL sections to prevent
linking to the default Docusaurus repository.