✨ feat(css): update font family to Fira Sans and adjust size
Changed the default font family from "Noto Sans" to "Fira Sans" for
base and heading styles in the CSS. Additionally, updated the base
font size from 16px to 17.5px to enhance readability. This change
improves the overall typography of the site.

Also added the Fira Sans font files to the project to support the
new font family.