✨ feat(articles): add "Comments in Rust" article and image
Created a new article titled "Comments in Rust" that provides an overview
of comments in the Rust programming language. This includes examples of
single-line and multi-line comments. Additionally, added a featured image
for the article to enhance visual appeal on the articles index page.

Updated the articles index to include a new InfoCard for easy navigation
to the "Comments in Rust" article.